QIDO-RS: Return Size of Study in private Attributes dcm4che/dcm4chee-arc-light#1051

+++ dcm4chee-arc-delete/src/main/java/org/dcm4chee/arc/delete/impl/DeletionServiceEJB.java
@@ -320,7 +320,6 @@ public class DeletionServiceEJB {
     }
 
     public void purgeInstanceRecordsOfSeries(Series.PkAndSize pkAndSize) {
-        HashMap<Long, UIDMap> uidMaps = new HashMap<>();
         List<Location> locations = em.createNamedQuery(Location.FIND_BY_SERIES_PK, Location.class)
                 .setParameter(1, pkAndSize.pk)
                 .getResultList();
@@ -330,25 +329,23 @@ public class DeletionServiceEJB {
         if (pkAndSize.size < 0)
             queryService.calculateSeriesSize(pkAndSize.pk);
         calculateMissingSeriesQueryAttributes(pkAndSize.pk);
-        Series series = locations.get(0).getInstance().getSeries();
         for (Location location : locations) {
-            switch (location.getObjectType()) {
-                case DICOM_FILE:
-                    UIDMap uidMap = location.getUidMap();
-                    if (uidMap != null)
-                        uidMaps.put(uidMap.getPk(), uidMap);
-                    em.remove(location);
-                    em.remove(location.getInstance());
-                    break;
-                case METADATA:
-                    storeEjb.removeOrMarkToDelete(location);
-                    break;
+            if (location.getObjectType() == Location.ObjectType.METADATA) {
+                location.setInstance(null);
+                location.setStatus(Location.Status.TO_DELETE);
             }
         }
-        for (UIDMap uidMap : uidMaps.values())
-            storeEjb.removeOrphaned(uidMap);
-        series.setInstancePurgeState(Series.InstancePurgeState.PURGED);
-        series.setInstancePurgeTime(null);
+        for (Location location : locations) {
+            if (location.getObjectType() == Location.ObjectType.DICOM_FILE) {
+                em.remove(location);
+                em.remove(location.getInstance());
+            }
+        }
+        em.createNamedQuery(Series.SET_INSTANCE_PURGE_STATE_AND_TIME)
+                .setParameter(1, pkAndSize.pk)
+                .setParameter(2, Series.InstancePurgeState.PURGED)
+                .setParameter(3, null)
+                .executeUpdate();
     }
 
     private void calculateMissingSeriesQueryAttributes(Long seriesPk) {
